Extract `GivenCommand` and add more tests for `RoleView`
