+++ users/src/test/java/io/spine/users/c/group/AssignRoleToGroupTest.java
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static io.spine.base.Identifier.newUuid;
-import static io.spine.users.c.group.given.GroupTestCommands.assignRoleToGroup;
+import static io.spine.users.given.GivenCommand.assignRoleToGroup;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
