+++ users/src/test/java/io/spine/users/c/group/UnassignRoleFromGroupTest.java
@@ -24,8 +24,8 @@ import io.spine.users.RoleId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static io.spine.users.c.group.given.GroupTestCommands.unassignRoleFromGroup;
 import static io.spine.users.c.group.given.GroupTestEnv.groupRole;
+import static io.spine.users.given.GivenCommand.unassignRoleFromGroup;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
