+++ users/src/test/java/io/spine/users/q/user/UserRolesProjectionIntegrationTest.java
@@ -40,17 +40,17 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static io.spine.testing.server.blackbox.verify.state.VerifyState.exactlyOne;
+import static io.spine.users.given.GivenCommand.assignRoleToGroup;
+import static io.spine.users.given.GivenCommand.assignRoleToUser;
+import static io.spine.users.given.GivenCommand.unassignRoleFromGroup;
+import static io.spine.users.given.GivenCommand.unassignRoleFromUser;
 import static io.spine.users.given.GivenId.groupUuid;
 import static io.spine.users.given.GivenId.roleUuid;
-import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.assignRoleToGroup;
-import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.assignRoleToUser;
 import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.createGroup;
 import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.createRole;
 import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.createUser;
 import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.joinGroup;
 import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.leaveGroup;
-import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.unassignRoleFromGroup;
-import static io.spine.users.q.user.given.UserRolesProjectionTestCommands.unassignRoleFromUser;
 import static io.spine.users.q.user.given.UserRolesProjectionTestEnv.roleDisplayName;
 import static io.spine.users.q.user.given.UserRolesProjectionTestEnv.userUuid;
 import static io.spine.users.q.user.given.UserRolesProjectionTestEnv.userWithRole;
